+++ packages/kit/test/apps/basics/src/routes/scroll/cross-document/a.svelte
@@ -1,0 +1,9 @@
+<script context="module">
+	export const router = false;
+</script>
+
+<h1>a</h1>
+
+<div style="height: 200vh; background: teal" />
+
+<a href="/scroll/cross-document/b">b</a>