+++ packages/kit/test/apps/basics/src/routes/scroll/cross-document/c.svelte
@@ -1,0 +1,5 @@
+<script context="module">
+	export const router = true;
+</script>
+
+<h1>c</h1>