+++ packages/kit/types/private.d.ts
@@ -134,12 +134,6 @@ export interface CspDirectives {
 	>;
 }
 
-export interface ErrorLoadInput<Params extends Record<string, string> = Record<string, string>>
-	extends LoadInput<Params> {
-	status?: number;
-	error?: Error;
-}
-
 export type HttpMethod = 'get' | 'head' | 'post' | 'put' | 'delete' | 'patch';
 
 export interface JSONObject {
@@ -155,28 +149,6 @@ export type JSONValue =
 	| ToJSON
 	| JSONValue[]
 	| JSONObject;
-
-export interface LoadInput<
-	Params extends Record<string, string> = Record<string, string>,
-	Props extends Record<string, any> = Record<string, any>
-> {
-	fetch(info: RequestInfo, init?: RequestInit): Promise<Response>;
-	params: Params;
-	props: Props;
-	routeId: string | null;
-	session: App.Session;
-	stuff: Partial<App.Stuff>;
-	url: URL;
-}
-
-export interface LoadOutput<Props extends Record<string, any> = Record<string, any>> {
-	status?: number;
-	error?: string | Error;
-	redirect?: string;
-	props?: Props;
-	stuff?: Partial<App.Stuff>;
-	maxage?: number;
-}
 
 export interface Logger {
 	(msg: string): void;